requirements.txt
acme>=0.31.0
certbot>=0.31.0
requests>=2.20.0
setuptools
zope.interface